lmapd: workspace: handle the schedule incoming queue
Add a function to move .meta and .data files (and only when the
pair is present) from the _incoming directory, to the "processing"
directory of a schedule, and call it before executing a schedule.

Add a function to remove files from the "processing" directory,
and call it when all actions of a schedule exit with a successful
status.  Do not remove directories (action workspaces), or anything
prefixed by "_".

CEPTRO.br-issue: MDC-473