lmapd: xml-io: use correct constants for action.state
The correct constants for action.state are LMAP_ACTION_STATE_*, not
LMAP_SCHEDULE_STATE_*.

Fortunately, these constants are numerically equivalent, so this did not
result in any runtime bug.

This issue propagated to json-io.c when it was expanted to match
capabilities with xml-io.c, and also needs to be fixed there.