lmapd: try a bit harder to detect stale pidfiles
Ignore illegal PIDs in a pidfile (negative or zero), and check if a
process with the PID read from the pidfile actually exists before
considering it valid.

This simple check can't handle the case when PID rollover (or a reboot
on a system where pidfiles are persistent) caused a process to have the
same PID as the PID in the stale pidfile, but it is far better than no
check at all.

Without this, lmapd will refuse to restart (until something deletes the
stale pidfile) if it ever gets killed by a signal (e.g.  SIGBUS due to a
NULL pointer derreference).  The simple check suffices to avoid this as
long as process churn is not so high as to cause PID rollover, and
system misconfiguration does not cause undue persistence of stale
pidfiles across reboots.

It also improves "lmapctl running" a bit, subject to the same caveats
described above.

A proper fix requires a full rework of the pidfile support, and likely
requires the use of flock() and fcntl(FD_CLOEXEC) since the pidfile must
be kept open and locked by the main [possibly daemonized] lmapd process
*only*, and closed in all other threads/children.

As for lmapctl, a proper fix would be to refactor its communication with
lmapd to use UNIX sockets instead of signal-driven update of data files.
Again, that'd be a very large change.