lmapd: workspace: protect private namespace for actions too
Skip the private namespace (directories and files prefixed by "_") when
moving action results and cleaning action workspaces.  This gives
actions a way to preserve state across separate runs of a schedule.

Hidden files are also skipped when moving an action's result to
destinations, but they are not preserved (they will be removed by
lmapd_workspace_action_clean()).

Note that actions cannot depend on the workspaces not being fully
cleaned up when lmapd is restarted or reconfigured, as such workspaces
depend on the lmapd task and schedule configuration.

CEPTRO.br-issue: MDC-1018
CEPTRO.br-issue: MDC-473