lmapd: workspace: restrict a schedule's active queue to regular files
Ensure the inodes we try to move from the incoming to the active queue
in a schedule workspace are in fact regular files.

The current workspace code and its filesystem layout is not prepared to
cope with subdirectories as queue items, and we don't want other inode
types getting moved/copied, either.

While at it, issue an error message when we fail to link() the "data"
file, and correct the error message when we fail to link() the "meta"
file (it was displaying the name of the "data" file).

CEPTRO.br-issue: MDC-473